requirements.txt
python-pushover>=0.2

--index-url https://pypi.python.org/simple/

-e .